requirements.txt
# Core ML (install torch separately with CUDA support)
# torch and torchvision installed via --index-url
numpy==1.26.4

# CLIP and transformers
transformers==4.35.0
sentence-transformers==2.2.2

# Image processing
Pillow==10.0.0
ftfy==6.1.3

# Optional: Stable Diffusion
diffusers==0.25.0

# Utils
tqdm==4.66.0
